grande progresso

+++ backend/src/Controllers/Category/DeleteCategory.ts
@@ -3,26 +3,35 @@ import prisma from "../../database";
 
 class DeleteCategory {
   async handle(req: Request, res: Response) {
-    const {id} = req.params;
+    const { id } = req.params;
 
     try {
+      if (!id) {
+        return res.status(401).json({ message: "error delete category line" });
+      }
 
-        if (!id) {
-        return res.status(401).json({ message: "error edit Category line" });
+      const categoryInUse = await prisma.product.findFirst({
+        where: { idCategory: id },
+      });
+
+      if (categoryInUse) {
+        return res.status(400).json({
+          message:
+            "Não é possível excluir: esta categoria está em uso em produtos.",
+        });
       }
 
       const deleteCategory = await prisma.category.delete({
-        where: { id: id }
+        where: { id: id },
       });
 
-      res
+      return res
         .status(200)
         .json({ message: "Delete Category Line: ", deleteCategory });
-
     } catch (error) {
-      res.status(500).json({ message: "Internal server error" });
+      return res.status(500).json({ message: "Internal server error" });
     }
   }
 }
 
-export {DeleteCategory};
+export { DeleteCategory };
